Update main Boarding Calendar.  JButtons take html code which has the ability to contain multiple lines of text in the button label.  This will be utilized when listing the pets that are being boarding on a specific day.
NOTE: This calendar is commented out from the HomeScreen class.  Since
it is dynamically created, it will break WindowBuilder if trying to edit
the UI.  To run the program with this calendar, uncomment the necessary
lines for adding the calendarViewBoarding to gl_panelBoarding.

Add Notes section for creating a Boarding Session.  This will be used
for adding instructions such as dietary needs for pets.